fix(sonar): resolve 482 SonarQube issues across codebase
- Delete CoverageGapTests.cs and redistribute 35 tests to appropriate files
- Fix redundant nameof() arguments in ArgumentNullException.ThrowIfNull calls
- Fix default(TKey) comparisons in InMemory/Entity/Mongo repositories
- Fix default returns in HttpUserAccessor generic methods
- Fix BLOCKER issues: constructor overload defaults, IDisposable pattern
- Fix ValueTask misuse in EntityManager.DisposeAsync
- Fix nested if statement in RepositoryLifecycleService (S1066)
- Fix getters not referencing backing fields in MultiTenant test suites
- Add pragma suppressions for intentional unused type parameters (S2326)
- Fix missing Expression import in RepositoryExtensionsSyncTests
