+++ test/Kista.Testing.TUnit/Suites/MultiTenantRepositoryTestSuite_T3.cs
@@ -25,7 +25,9 @@ public abstract class MultiTenantRepositoryTestSuite<TTenantInfo, TPerson, TKey>
 
     protected virtual int EntitySetCount => 100;
 
-    protected virtual string[] TenantIds => new[] { "tenant1", "tenant2", "tenant3" };
+    private static readonly string[] DefaultTenantIds = { "tenant1", "tenant2", "tenant3" };
+
+    protected virtual string[] TenantIds => DefaultTenantIds;
 
     protected virtual IDictionary<string, IReadOnlyList<TPerson>>? People { get; private set; }
